Moved libtiff test into libtiff test file
